+++ client/app/bundles/course/assessment/question/programming/components/BuildLog.jsx
@@ -1,8 +1,25 @@
 import Immutable from 'immutable';
 
 import React, { PropTypes } from 'react';
-import { injectIntl } from 'react-intl';
-import { buildLogTranslations as translations } from '../constants/translations';
+import { injectIntl, defineMessages } from 'react-intl';
+
+const translations = defineMessages({
+  header: {
+    id: 'course.assessment.question.programming.buildLog.header',
+    defaultMessage: 'Build Log',
+    description: 'Header for build log.',
+  },
+  stdoutHeader: {
+    id: 'course.assessment.question.programming.buildLog.stdoutHeader',
+    defaultMessage: 'Standard Output',
+    description: 'Header for standard output.',
+  },
+  stderrHeader: {
+    id: 'course.assessment.question.programming.buildLog.stderrHeader',
+    defaultMessage: 'Standard Error',
+    description: 'Header for standard error.',
+  },
+});
 
 const propTypes = {
   buildLogData: PropTypes.instanceOf(Immutable.Map).isRequired,
