+++ client/app/bundles/course/assessment/question/programming/components/OnlineEditorPythonView.intl.js
@@ -1,0 +1,68 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  autograded: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.autograded',
+    defaultMessage: 'Autograded',
+  },
+  prependTitle: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.prependTitle',
+    defaultMessage: 'Prepend',
+    description: 'Title for prepend code block.',
+  },
+  appendTitle: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.appendTitle',
+    defaultMessage: 'Append',
+    description: 'Title for append code block.',
+  },
+  solutionTitle: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.solutionTitle',
+    defaultMessage: 'Solution Template',
+    description: 'Title for solution template code block.',
+  },
+  submissionTitle: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.submissionTitle',
+    defaultMessage: 'Submission Template',
+    description: 'Title for submission template code block.',
+  },
+  addNewTestButton: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.addNewTestButton',
+    defaultMessage: 'Add new test',
+    description: 'Button for adding new test case.',
+  },
+  publicTestCases: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.publicTestCases',
+    defaultMessage: 'Public Test Cases',
+    description: 'Title for public test cases panel.',
+  },
+  privateTestCases: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.privateTestCases',
+    defaultMessage: 'Private Test Cases',
+    description: 'Title for private test cases panel.',
+  },
+  evaluationTestCases: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.evaluationTestCases',
+    defaultMessage: 'Evaluation Test Cases',
+    description: 'Title for evaluation test cases panel.',
+  },
+  identifierHeader: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.identifierHeader',
+    defaultMessage: 'Identifier',
+    description: 'Header for identifier column of test cases panel.',
+  },
+  expressionHeader: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.expressionHeader',
+    defaultMessage: 'Expression',
+    description: 'Header for expression column of test cases panel.',
+  },
+  expectedHeader: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.expectedHeader',
+    defaultMessage: 'Expected',
+    description: 'Header for expected column of test cases panel.',
+  },
+  hintHeader: {
+    id: 'course.assessment.question.programming.onlineEditorPythonView.hintHeader',
+    defaultMessage: 'Hint',
+    description: 'Header for hint column of test cases panel.',
+  },
+});
