+++ client/app/bundles/course/assessment/question/programming/components/ProgrammingQuestionForm.intl.js
@@ -1,0 +1,89 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  titleFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.titleFieldLabel',
+    defaultMessage: 'Title',
+    description: 'Label for the title input field.',
+  },
+  descriptionFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.descriptionFieldLabel',
+    defaultMessage: 'Description',
+    description: 'Label for the description input field.',
+  },
+  staffOnlyCommentsFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.staffOnlyCommentsFieldLabel',
+    defaultMessage: 'Staff only comments',
+    description: 'Label for the staff only comments input field.',
+  },
+  maximumGradeFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.maximumGradeFieldLabel',
+    defaultMessage: 'Maximum Grade',
+    description: 'Label for the maximum grade input field.',
+  },
+  skillsFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.skillsFieldLabel',
+    defaultMessage: 'Skills',
+    description: 'Label for the skills input field.',
+  },
+  languageFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.languageFieldLabel',
+    defaultMessage: 'Language',
+    description: 'Label for the language input field.',
+  },
+  memoryLimitFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.memoryLimitFieldLabel',
+    defaultMessage: 'Memory Limit',
+    description: 'Label for the memory limit input field.',
+  },
+  timeLimitFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.timeLimitFieldLabel',
+    defaultMessage: 'Time Limit',
+    description: 'Label for the time limit input field.',
+  },
+  attemptLimitFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.attemptLimitFieldLabel',
+    defaultMessage: 'Attempt Limit',
+    description: 'Label for the attempt limit input field.',
+  },
+  templatePackageFieldLabel: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.templatePackageFieldLabel',
+    defaultMessage: 'Template Package',
+    description: 'Label for the template package input field.',
+  },
+  attemptLimitPlaceholderMessage: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.attemptLimitPlaceholderMessage',
+    defaultMessage: 'The maximum times that the students can test their answers (does not apply to staff)',
+    description: 'Placeholder message for attempt limit input field.',
+  },
+  downloadPackage: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.downloadPackage',
+    defaultMessage: 'Download Package',
+    description: 'Label for downloading the zip package.',
+  },
+  editTestsOnlineButton: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.editTestsOnlineButton',
+    defaultMessage: 'Use Test Case Editor',
+    description: 'Button for editing tests online.',
+  },
+  uploadPackageButton: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.uploadPackageButton',
+    defaultMessage: 'Upload Package',
+    description: 'Button for uploading package.',
+  },
+  submitButton: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.submitButton',
+    defaultMessage: 'Submit',
+    description: 'Button for submitting the form.',
+  },
+  evaluatingMessage: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.evaluatingMessage',
+    defaultMessage: 'Evaluating',
+    description: 'Text to be displayed when evaluating the programming question.',
+  },
+  loadingMessage: {
+    id: 'course.assessment.question.programming.programmingQuestionForm.loadingMessage',
+    defaultMessage: 'Loading',
+    description: 'Text to be displayed when waiting for server response after form submission.',
+  },
+});
